Moved code to a LoadRegion class
